general/panorama/requirements.txt
# PDF report (professional layout, pie charts, tables). Required for output_formats=pdf.
reportlab>=4.0.0
# Excel report (FRONTPAGE, SAST, SCA, LICENSES sheets). Required for output_formats=xlsx.
openpyxl>=3.1.0
# Infra (Dockerfile, compose, K8s YAML parsing).
PyYAML>=6.0
